+++ cascade-oj/ent/announcement/where.go
@@ -10,52 +10,52 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int) predicate.Announcement {
+func ID(id int64) predicate.Announcement {
 	return predicate.Announcement(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int) predicate.Announcement {
+func IDEQ(id int64) predicate.Announcement {
 	return predicate.Announcement(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int) predicate.Announcement {
+func IDNEQ(id int64) predicate.Announcement {
 	return predicate.Announcement(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int) predicate.Announcement {
+func IDIn(ids ...int64) predicate.Announcement {
 	return predicate.Announcement(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int) predicate.Announcement {
+func IDNotIn(ids ...int64) predicate.Announcement {
 	return predicate.Announcement(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int) predicate.Announcement {
+func IDGT(id int64) predicate.Announcement {
 	return predicate.Announcement(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int) predicate.Announcement {
+func IDGTE(id int64) predicate.Announcement {
 	return predicate.Announcement(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int) predicate.Announcement {
+func IDLT(id int64) predicate.Announcement {
 	return predicate.Announcement(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int) predicate.Announcement {
+func IDLTE(id int64) predicate.Announcement {
 	return predicate.Announcement(sql.FieldLTE(FieldID, id))
 }
 
 // PublisherID applies equality check predicate on the "publisher_id" field. It's identical to PublisherIDEQ.
-func PublisherID(v int) predicate.Announcement {
+func PublisherID(v int64) predicate.Announcement {
 	return predicate.Announcement(sql.FieldEQ(FieldPublisherID, v))
 }
 
@@ -70,22 +70,22 @@ func Content(v string) predicate.Announcement {
 }
 
 // PublisherIDEQ applies the EQ predicate on the "publisher_id" field.
-func PublisherIDEQ(v int) predicate.Announcement {
+func PublisherIDEQ(v int64) predicate.Announcement {
 	return predicate.Announcement(sql.FieldEQ(FieldPublisherID, v))
 }
 
 // PublisherIDNEQ applies the NEQ predicate on the "publisher_id" field.
-func PublisherIDNEQ(v int) predicate.Announcement {
+func PublisherIDNEQ(v int64) predicate.Announcement {
 	return predicate.Announcement(sql.FieldNEQ(FieldPublisherID, v))
 }
 
 // PublisherIDIn applies the In predicate on the "publisher_id" field.
-func PublisherIDIn(vs ...int) predicate.Announcement {
+func PublisherIDIn(vs ...int64) predicate.Announcement {
 	return predicate.Announcement(sql.FieldIn(FieldPublisherID, vs...))
 }
 
 // PublisherIDNotIn applies the NotIn predicate on the "publisher_id" field.
-func PublisherIDNotIn(vs ...int) predicate.Announcement {
+func PublisherIDNotIn(vs ...int64) predicate.Announcement {
 	return predicate.Announcement(sql.FieldNotIn(FieldPublisherID, vs...))
 }
