+++ cascade-oj/ent/systemlog/where.go
@@ -10,47 +10,47 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int) predicate.SystemLog {
+func ID(id int64) predicate.SystemLog {
 	return predicate.SystemLog(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int) predicate.SystemLog {
+func IDEQ(id int64) predicate.SystemLog {
 	return predicate.SystemLog(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int) predicate.SystemLog {
+func IDNEQ(id int64) predicate.SystemLog {
 	return predicate.SystemLog(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int) predicate.SystemLog {
+func IDIn(ids ...int64) predicate.SystemLog {
 	return predicate.SystemLog(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int) predicate.SystemLog {
+func IDNotIn(ids ...int64) predicate.SystemLog {
 	return predicate.SystemLog(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int) predicate.SystemLog {
+func IDGT(id int64) predicate.SystemLog {
 	return predicate.SystemLog(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int) predicate.SystemLog {
+func IDGTE(id int64) predicate.SystemLog {
 	return predicate.SystemLog(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int) predicate.SystemLog {
+func IDLT(id int64) predicate.SystemLog {
 	return predicate.SystemLog(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int) predicate.SystemLog {
+func IDLTE(id int64) predicate.SystemLog {
 	return predicate.SystemLog(sql.FieldLTE(FieldID, id))
 }
